Separate Lambda function logic part into handler module
