+++ src/07-02-todo-final/serverless/models/model.js
@@ -1,4 +1,4 @@
-import mongoose from '../src/database';
+import mongoose from './database';
 
 const { Schema } = mongoose;
